correct 102-fibonacci.c - change to ld because lld not supported

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -9,8 +9,8 @@
 int main(void)
 {
 	int i;
-	long long int j, k;
-	long long int numb[i];
+	long int j, k;
+	long int numb[i];
 
 	numb[0] = 0;
 	numb[1] = 1;
@@ -24,7 +24,7 @@ int main(void)
 		{
 			printf(", ");
 		}
-		printf("%lld", numb[i]);
+		printf("%ld", numb[i]);
 	}
 
 	printf("\n");
